Moved extension into correct file and removed unrequited file.
