--- Say Their Names/Source/Utils/Extentions/UICollectionView+Extentions.swift
@@ -1,34 +1,0 @@
-//
-//  UICollectionView+Extentions.swift
-//  Say Their Names
-//
-//  Created by Thomas Murray on 03/06/2020.
-//  Copyright © 2020 Franck-Stephane Ndame Mpouli. All rights reserved.
-//
-
-import UIKit
-
-extension UICollectionView {
-    func scrollToNearestVisibleCollectionViewCell() {
-        self.decelerationRate = UIScrollView.DecelerationRate.fast
-        let visibleCenterPositionOfScrollView = Float(self.contentOffset.x + (self.bounds.size.width / 2))
-        var closestCellIndex = -1
-        var closestDistance: Float = .greatestFiniteMagnitude
-        for i in 0..<self.visibleCells.count {
-            let cell = self.visibleCells[i]
-            let cellWidth = cell.bounds.size.width
-            let cellCenter = Float(cell.frame.origin.x + cellWidth / 2)
-
-            // Now calculate closest cell
-            let distance: Float = fabsf(visibleCenterPositionOfScrollView - cellCenter)
-            if distance < closestDistance {
-                closestDistance = distance
-                closestCellIndex = self.indexPath(for: cell)!.row
-            }
-        }
-        if closestCellIndex != -1 {
-            self.scrollToItem(at: IndexPath(row: closestCellIndex, section: 0), at: .centeredHorizontally, animated: true)
-        }
-    }
-}
-
